+++ PayPalWebPayments/src/main/java/com/paypal/android/paypalwebpayments/PayPalWebVaultRequest.kt
@@ -7,23 +7,8 @@ import com.paypal.android.corepayments.ReturnToAppStrategy
  *
  * @property [setupTokenId] ID for the setup token associated with the vault approval
  * @property [returnToAppStrategy] Strategy for returning to the app after checkout flow
- * @property [approveVaultHref] URL for the approval web page
  */
-data class PayPalWebVaultRequest @Deprecated("Use PayPalWebVaultRequest(setupTokenId) instead.")
-constructor(
+data class PayPalWebVaultRequest(
     val setupTokenId: String,
-    val returnToAppStrategy: ReturnToAppStrategy? = null,
-    @Deprecated("The approveVaultHref property is no longer required and will be ignored.")
-    val approveVaultHref: String? = null // NEXT_MAJOR_VERSION: - Remove this property
-) {
-
-    /**
-     * Request to vault a PayPal payment method using [PayPalWebCheckoutClient.vault].
-     *
-     * @property [setupTokenId] ID for the setup token associated with the vault approval
-     */
-    constructor(
-        setupTokenId: String,
-        returnToAppStrategy: ReturnToAppStrategy? = null
-    ) : this(setupTokenId, returnToAppStrategy, null)
-}
+    val returnToAppStrategy: ReturnToAppStrategy? = null
+)